split lib.rs into modules, optimize case normalizer by precomputing some info about target string
